make ping() properly async
The current testConnection() uses a synchronous native API which means
it completely blocks the event loop until it finishes. Normally this is
somewhat expected, but if the connection is being black-holed or
something similar then the request will need to timeout - synchronously.
This means the event loop is blocked for many minutes, preventing even a
setTimeout() from being triggered during that time.